requirements.txt
annotated-types==0.7.0
anyio==4.8.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
fastapi==0.115.7
h11==0.14.0
idna==3.10
joblib==1.4.2
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
platformdirs==4.3.6
pooch==1.8.2
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-multipart==0.0.20
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
starlette==0.45.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
tensorflow
tensorflow_hub
Pillow
pandas
flask
